+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -180,6 +180,7 @@ my $expected_tested_by = [
     't/1280-isf-enum-member-inline-drive-expression-values.t',
     't/1281-isf-enum-member-library-use-params.t',
     't/1282-isf-enum-member-drive-expression-values.t',
+    't/1283-isf-aggregate-rule-values.t',
 ];
 
 subtest 'direct ISF tested_by metadata is exact and valid' => sub {